+++ app/Http/Controllers/taskController.php
@@ -20,6 +20,8 @@ class taskController extends Controller
      */
     public function index( $project_id = null )
     {
+        if( !have_permission('','','can_view_tasks') ) return view('admin.access_error');
+
         if(  $project_id ) {
             $tasks = Task::where('project_id',$project_id)->get();
         } else {
@@ -30,12 +32,18 @@ class taskController extends Controller
     }
 
     public function module_tasks( $module_id ) {
-        $tasks = Tasks::where('module_id',$module_id)->get();
+
+        if( !have_permission('','','can_view_tasks') ) return view('admin.access_error');
+
+        $tasks = Task::where('module_id',$module_id)->get();
         return view( 'admin.task.index',compact('tasks') );
     }
 
     public function tasklist_tasks( $tasklist_id ) {
-        $tasks = Tasks::where('tasklist_id',$tasklist_id)->get();
+
+        if( !have_permission('','','can_view_tasks') ) return view('admin.access_error');
+
+        $tasks = Task::where('tasklist_id',$tasklist_id)->get();
         return view( 'admin.task.index',compact('tasks') );
     }
 
@@ -48,6 +56,8 @@ class taskController extends Controller
      */
     public function create( $project_id = null , $module_id = null , $tasklist_id = null )
     {
+        if( !have_permission('','','can_create_tasks') ) return view('admin.access_error');
+
         $statuses = common_stuff::get_status_options();
 
         $projects = array('') + Project::lists('title','id')->toArray();
@@ -65,6 +75,9 @@ class taskController extends Controller
      * Create by module
      */
     public function create_by_module( $module_id = null ){
+
+        if( !have_permission('','','can_create_tasks') ) return view('admin.access_error');
+
         $project_id = Module::where('id',$module_id)->pluck('project_id')[0];
         return $this->create( $project_id, $module_id );
     }
@@ -73,6 +86,9 @@ class taskController extends Controller
      * Create by tasklist
      */
     public function create_by_tasklist( $tasklist_id ) {
+
+        if( !have_permission('','','can_create_tasks') ) return view('admin.access_error');
+
         $test = Tasklist::where('id',$tasklist_id)->pluck('project_id','module_id');
         $module_id = key($test->toArray());
         $project_id = $test->toArray()[$module_id];
@@ -87,6 +103,8 @@ class taskController extends Controller
      */
     public function store(Request $request)
     {
+        if( !have_permission('','','can_create_tasks') ) return view('admin.access_error');
+
         $task = Task::create($request->all());
         $task->assigned_users()->sync($request->user_id);
         $task->user()->associate(get_current_user_id());
@@ -103,6 +121,8 @@ class taskController extends Controller
      */
     public function show($id)
     {
+        if( !have_permission('','','can_view_task') ) return view('admin.access_error');
+
         $task = Task::find($id);
         return view('admin.task.single',compact('task'));
     }
@@ -115,6 +135,8 @@ class taskController extends Controller
      */
     public function edit($id)
     {
+        if( !have_permission('','','can_edit_tasks') ) return view('admin.access_error');
+
         $task = Task::find($id);
         $task->assigned_users = json_decode($task->assigned_users);
         $task->assigned_users = array_map(function($item) {
@@ -141,6 +163,8 @@ class taskController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if( !have_permission('','','can_edit_tasks') ) return view('admin.access_error');
+
         $task = Task::find($id);
         $task->update($request->all());
         $task->assigned_users()->sync($request->user_id);
@@ -158,6 +182,8 @@ class taskController extends Controller
      */
     public function destroy($id)
     {
+        if( !have_permission('','','can_delete_tasks') ) return view('admin.access_error');
+
         Task::destroy($id);
         return redirect()->route('admin.tasks.index');
     }
